requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.30
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.2.1
